Add files via upload
You are tasked with creating a simple number analysis tool for a basic application. This tool
needs to perform operations on a list of numbers provided by the user. Specifically, you need
to compute the sum, average, and count of even and odd numbers from the list. You will use
loops to process these numbers.
Tasks:
1. Create a C++ program that:
• Prompts the user to enter the number of values they want to input.
• Accepts that many numbers from the user.
• Uses loops to calculate the following:
• The sum of all entered numbers.
• The average of all entered numbers.
• The count of even numbers.
• The count of odd numbers.
• Displays the results.